Fixed the looking glass slot state transition issue.
There is a possibility that the app crash signal of previous
opened app will be sent to the newly opened app wrongly just
before the new looking glass window popup. This caused incorrect
looking glass slot allocation state in daemon and users could
open more than one looking glass window with same app content.

Tracked-On: OAM-101245

+++ host/src/pack/dpkg/opt/cfc/mwc/bin/kill_lg_instance.sh
@@ -18,7 +18,7 @@ else
    pid=`ps aux | grep "LG_B1_Client " | grep -v grep | grep $inst | awk '{print $2}'`
 fi
 
-if [ "$pid" -gt 0 ]
+if [ ! -z "$pid" ] && [ "$pid" -gt 0 ]
 then
     echo  "inside kill"	
     kill -KILL $pid
